+++ schemas/hardware_vtep/src/main/java/org/opendaylight/ovsdb/schema/hardwarevtep/Manager.java
@@ -30,16 +30,16 @@ public interface Manager extends TypedBaseTable<GenericTableSchema> {
     public void setTarget(String target);
 
     @TypedColumn(name="max_backoff", method=MethodType.GETCOLUMN, fromVersion="1.0.0")
-    public Column<GenericTableSchema, Set<Integer>> getMaxBackoffColumn();
+    public Column<GenericTableSchema, Set<Long>> getMaxBackoffColumn();
 
     @TypedColumn(name="max_backoff", method=MethodType.SETDATA, fromVersion="1.0.0")
-    public void setMaxBackoff(Set<Integer> maxBackoff);
+    public void setMaxBackoff(Set<Long> maxBackoff);
 
     @TypedColumn(name="inactivity_probe", method=MethodType.GETCOLUMN, fromVersion="1.0.0")
-    public Column<GenericTableSchema, Set<Integer>> getInactivityProbeColumn();
+    public Column<GenericTableSchema, Set<Long>> getInactivityProbeColumn();
 
     @TypedColumn(name="inactivity_probe", method=MethodType.SETDATA, fromVersion="1.0.0")
-    public void setInactivityProbe(Set<Integer> inactivityProbe);
+    public void setInactivityProbe(Set<Long> inactivityProbe);
 
     @TypedColumn(name="other_config", method=MethodType.GETCOLUMN, fromVersion="1.0.0")
     public Column<GenericTableSchema, Map<String, String>> getOtherConfigColumn();
